requirements.txt
opencv-python
numpy
Pillow
sentinelhub
